% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{AbsolutePercentageError}
\alias{AbsolutePercentageError}
\title{Calculates APE: https://blog.arkieva.com/two-sides-of-the-mape-coin/}
\usage{
AbsolutePercentageError(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
APE
}
\description{
Calculates APE: https://blog.arkieva.com/two-sides-of-the-mape-coin/
}
