% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{evaluateClassification}
\alias{evaluateClassification}
\title{Calculates various classification metrics}
\usage{
evaluateClassification(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
data frame with classification metrics
}
\description{
Calculates various classification metrics
}
