% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{MeanAbsolutePercentageError}
\alias{MeanAbsolutePercentageError}
\title{Calculates MAPE: https://en.wikipedia.org/wiki/Mean_absolute_percentage_error}
\usage{
MeanAbsolutePercentageError(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
MAPE
}
\description{
Calculates MAPE: https://en.wikipedia.org/wiki/Mean_absolute_percentage_error
}
