% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{MedianAbsolutePercentageError}
\alias{MedianAbsolutePercentageError}
\title{Calculates median APE}
\usage{
MedianAbsolutePercentageError(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
Median APE
}
\description{
Calculates median APE
}
