% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{metricAccuracy}
\alias{metricAccuracy}
\title{Calculates classification accuracy}
\usage{
metricAccuracy(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
Classification accuracy
}
\description{
Calculates classification accuracy
}
