% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{metricFscore}
\alias{metricFscore}
\title{Calculates classification F-score}
\usage{
metricFscore(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
Classification F-score
}
\description{
Calculates classification F-score
}
