% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{metricKappa}
\alias{metricKappa}
\title{Calculates Kappa metric for classification}
\usage{
metricKappa(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
Kappa metric
}
\description{
Calculates Kappa metric for classification
}
