% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{metricPrecision}
\alias{metricPrecision}
\title{Calculates classification precision}
\usage{
metricPrecision(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
Classification precision
}
\description{
Calculates classification precision
}
