% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{metricSensitivity}
\alias{metricSensitivity}
\title{Calculates classification sensitivity}
\usage{
metricSensitivity(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
Classification sensitivity
}
\description{
Calculates classification sensitivity
}
