% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Radjsquared}
\alias{Radjsquared}
\title{Calculates adjusted R-squared metric}
\usage{
Radjsquared(predicted, actual, N, nvar)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
Adjusted R-squared metric
}
\description{
Calculates adjusted R-squared metric
}
