% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{RootMSE}
\alias{RootMSE}
\title{Calculates root mean square error}
\usage{
RootMSE(v1, v2)
}
\arguments{
\item{v1}{first vector of values}

\item{v2}{second vector of values}
}
\value{
root mean square error
}
\description{
Calculates root mean square error
}
