% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Rsquared}
\alias{Rsquared}
\title{Calculates R-squared metric}
\usage{
Rsquared(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
R-squared metric
}
\description{
Calculates R-squared metric
}
