% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{SymmetricMeanAbsolutePercentageError}
\alias{SymmetricMeanAbsolutePercentageError}
\title{Calculates SMAPE: https://en.wikipedia.org/wiki/Symmetric_mean_absolute_percentage_error}
\usage{
SymmetricMeanAbsolutePercentageError(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
SMAPE
}
\description{
Calculates SMAPE: https://en.wikipedia.org/wiki/Symmetric_mean_absolute_percentage_error
}
